This citation is now for the actual course rather than the textbook
I am updating the citation to reflect those who contributed to developing course content. Lauren was a reviewer for the textbook but did not contribute to content for the actual course. this is dated because we used to include all lessons with the course structure and content and now they are in a separate textbook. Lauren is listed as a contributor in that textbook.